requirements.txt
playwright==1.42.0
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.31.0
pytest==8.0.2
pytest-playwright==0.4.4 